[refactor, fix] khushal87 LinearProgress, PricingCard, SocialIcon, Tile
