+++ src/PricingCard/__tests__/PricingCard.tsx
@@ -4,7 +4,8 @@ import toJson from 'enzyme-to-json';
 import { create } from 'react-test-renderer';
 import theme from '../../config/theme';
 import { ThemeProvider } from '../../config';
-import ThemedPricingCard, { PricingCard, PricingButton } from '../index';
+import ThemedPricingCard, { PricingCard } from '../index';
+import { PricingButton } from '../components/PricingButton';
 import Icon from '../../icons/Icon';
 import Button from '../../Button';
