+++ src/PricingCard/index.tsx
@@ -1,7 +1,6 @@
 import { withTheme } from '../config';
 import { PricingCard, PricingCardProps } from './PricingCard';
-import { PricingButton } from './components/PricingButton';
 
-export { PricingCard, PricingButton };
+export { PricingCard };
 export type { PricingCardProps };
 export default withTheme(PricingCard, 'PricingCard');
