+++ src/Tile/index.tsx
@@ -1,7 +1,6 @@
 import { withTheme } from '../config';
 import { Tile, TileProps } from './Tile';
-import { FeaturedTile } from './components/FeaturedTile';
 
-export { Tile, FeaturedTile };
+export { Tile };
 export type { TileProps };
 export default withTheme(Tile, 'Tile');
